feat: separated hooks implementation
