+++ src/hooks/useCharacters.ts
@@ -1,6 +1,6 @@
 import { useCharacterStore } from '@/store/characterStore';
 import { useQuery } from '@tanstack/vue-query';
-import { getCharacters, getSingleCharacter, getMultipleCharacters } from '../api/characters';
+import { getCharacters } from '../api/characters';
 import { storeToRefs } from 'pinia';
 import { watchEffect } from 'vue';
 
@@ -24,47 +24,5 @@ export const useCharacters = () => {
     currentPage,
     totalPages,
     getPage: store.setPageCharacter
-  };
-};
-
-export const useCurrentCharacter = (characterId: number) => {
-  const store = useCharacterStore();
-  const { currentCharacter } = storeToRefs(store);
-
-  const { data, isFetching, isError } = useQuery({
-    queryKey: ['character'],
-    queryFn: () => getSingleCharacter(characterId),
-    retry: false
-  });
-
-  watchEffect(() => {
-    if (data.value) store.setCurrentCharacter(data.value);
-  });
-
-  return {
-    currentCharacter,
-    isFetching,
-    isError
-  };
-};
-
-export const useLatestCharacters = () => {
-  const store = useCharacterStore();
-  const { latestCharacters } = storeToRefs(store);
-
-  const { data, isFetching, isError } = useQuery({
-    queryKey: ['character'],
-    queryFn: () => getMultipleCharacters('2,4'),
-    retry: false
-  });
-
-  watchEffect(() => {
-    if (data.value) store.setLatestCharacters(data.value);
-  });
-
-  return {
-    latestCharacters,
-    isFetching,
-    isError
   };
 };
